+++ backend/canisters/community/impl/src/updates/c2c_set_user_suspended.rs
@@ -16,7 +16,7 @@ fn c2c_set_user_suspended_impl(args: Args, state: &mut RuntimeState) -> Response
     let now = state.env.now();
     if state.data.members.set_suspended(args.user_id, args.suspended, now).is_some() {
         for channel_id in state.data.members.channels_for_member(args.user_id) {
-            if let Some(channel) = state.data.channels.get_mut(&channel_id) {
+            if let Some(channel) = state.data.channels.get_mut(channel_id) {
                 channel.chat.members.set_suspended(args.user_id, args.suspended, now);
             }
         }