+++ backend/integration_tests/src/communities/delete_channel_tests.rs
@@ -65,9 +65,7 @@ fn stable_memory_garbage_collected_after_deleting_channel() {
         ..
     } = init_test_data(env, canister_ids, *controller);
 
-    let stable_map = get_stable_memory_map(env, community_id, STABLE_MEMORY_MAP_MEMORY_ID);
-    // 11 keys = 1 community event, 2 community members, 2 chat events and 2 users per channel
-    assert_eq!(stable_map.len(), 11);
+    let initial_stable_memory_map_keys = get_stable_memory_map(env, community_id, STABLE_MEMORY_MAP_MEMORY_ID).len();
 
     for _ in 0..100 {
         client::community::happy_path::send_text_message(env, &user1, community_id, channel_id1, None, random_string(), None);
@@ -75,7 +73,7 @@ fn stable_memory_garbage_collected_after_deleting_channel() {
 
     assert_eq!(
         get_stable_memory_map(env, community_id, STABLE_MEMORY_MAP_MEMORY_ID).len(),
-        111
+        initial_stable_memory_map_keys + 100
     );
 
     for _ in 0..80 {
@@ -84,7 +82,7 @@ fn stable_memory_garbage_collected_after_deleting_channel() {
 
     assert_eq!(
         get_stable_memory_map(env, community_id, STABLE_MEMORY_MAP_MEMORY_ID).len(),
-        191
+        initial_stable_memory_map_keys + 180
     );
 
     client::community::happy_path::delete_channel(env, user1.principal, community_id, channel_id1);
@@ -94,7 +92,7 @@ fn stable_memory_garbage_collected_after_deleting_channel() {
 
     assert_eq!(
         get_stable_memory_map(env, community_id, STABLE_MEMORY_MAP_MEMORY_ID).len(),
-        88
+        initial_stable_memory_map_keys + 77
     );
 
     client::community::happy_path::delete_channel(env, user1.principal, community_id, channel_id2);
@@ -102,7 +100,10 @@ fn stable_memory_garbage_collected_after_deleting_channel() {
     env.advance_time(Duration::from_secs(60));
     env.tick();
 
-    assert_eq!(get_stable_memory_map(env, community_id, STABLE_MEMORY_MAP_MEMORY_ID).len(), 5);
+    assert_eq!(
+        get_stable_memory_map(env, community_id, STABLE_MEMORY_MAP_MEMORY_ID).len(),
+        initial_stable_memory_map_keys - 6
+    );
 }
 
 fn init_test_data(env: &mut PocketIc, canister_ids: &CanisterIds, controller: Principal) -> TestData {